Updated password verification
Passwords are now hashed on the server side so that when a user
uninstall the application they do not loose their account, since the
salt is saved in the server.
Register: now hashes the password with salt and saves the output with
the salt to the db.
Sign in: checks if the user provided password hashes the same results
saved in the db after the same salt is applied.